Added some new images
