--- examples/src/main/kotlin/examples/onnx/poseestimation/MoveNetSinglePoseLighting.kt
@@ -1,181 +1,0 @@
-/*
- * Copyright 2020 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
- */
-
-package examples.onnx.poseestimation
-
-import examples.transferlearning.getFileFromResource
-import org.jetbrains.kotlinx.dl.api.inference.loaders.ONNXModelHub
-import org.jetbrains.kotlinx.dl.api.inference.onnx.ONNXModels
-import org.jetbrains.kotlinx.dl.dataset.image.ColorMode
-import org.jetbrains.kotlinx.dl.dataset.preprocessor.*
-import org.jetbrains.kotlinx.dl.dataset.preprocessor.image.convert
-import org.jetbrains.kotlinx.dl.dataset.preprocessor.image.resize
-import java.io.File
-
-/**
- * This examples demonstrates the inference concept on SSD model:
- * - Model is obtained from [ONNXModelHub].
- * - Model predicts on a few images located in resources.
- * - Special preprocessing is applied to images before prediction.
- */
-fun moveNetRawPrediction() {
-    val modelHub = ONNXModelHub(cacheDirectory = File("cache/pretrainedModels"))
-    val modelType = ONNXModels.PoseEstimation.MoveNetSinglePoseLighting
-    val model = modelHub.loadModel(modelType)
-
-    model.use {
-        println(it)
-
-        val imageFile = getFileFromResource("datasets/poses/1.jpg")
-        val preprocessing: Preprocessing = preprocess {
-            load {
-                pathToData = imageFile
-                imageShape = ImageShape(null, null, 3)
-            }
-            transformImage {
-                resize {
-                    outputHeight = 192
-                    outputWidth = 192
-                }
-                convert { colorMode = ColorMode.BGR }
-            }
-        }
-
-        val inputData = modelType.preprocessInput(preprocessing)
-
-        val yhat = it.predictRaw(inputData)
-        println(yhat.toTypedArray().contentDeepToString())
-        visualisePosePoints(imageFile, (yhat as List<Array<Array<Array<FloatArray>>>>)[0][0][0])
-        // Dictionary that maps from joint names to keypoint indices.
-        val keypoints = {
-            "nose" to 0
-            "left_eye" to 1
-            "right_eye" to 2
-            "left_ear" to 3
-            "right_ear" to 4
-            "left_shoulder" to 5
-            "right_shoulder" to 6
-            "left_elbow" to 7
-            "right_elbow" to 8
-            "left_wrist" to 9
-            "right_wrist" to 10
-            "left_hip" to 11
-            "right_hip" to 12
-            "left_knee" to 13
-            "right_knee" to 14
-            "left_ankle" to 15
-            "right_ankle" to 16
-        }
-
-    }
-}
-
-
-/*
-
-#@title Helper functions for visualization
-
-# Dictionary that maps from joint names to keypoint indices.
-KEYPOINT_DICT = {
-    'nose': 0,
-    'left_eye': 1,
-    'right_eye': 2,
-    'left_ear': 3,
-    'right_ear': 4,
-    'left_shoulder': 5,
-    'right_shoulder': 6,
-    'left_elbow': 7,
-    'right_elbow': 8,
-    'left_wrist': 9,
-    'right_wrist': 10,
-    'left_hip': 11,
-    'right_hip': 12,
-    'left_knee': 13,
-    'right_knee': 14,
-    'left_ankle': 15,
-    'right_ankle': 16
-}
-
-# Maps bones to a matplotlib color name.
-KEYPOINT_EDGE_INDS_TO_COLOR = {
-    (0, 1): 'm',
-    (0, 2): 'c',
-    (1, 3): 'm',
-    (2, 4): 'c',
-    (0, 5): 'm',
-    (0, 6): 'c',
-    (5, 7): 'm',
-    (7, 9): 'm',
-    (6, 8): 'c',
-    (8, 10): 'c',
-    (5, 6): 'y',
-    (5, 11): 'm',
-    (6, 12): 'c',
-    (11, 12): 'y',
-    (11, 13): 'm',
-    (13, 15): 'm',
-    (12, 14): 'c',
-    (14, 16): 'c'
-}
-
-def _keypoints_and_edges_for_display(keypoints_with_scores,
-                                     height,
-                                     width,
-                                     keypoint_threshold=0.11):
-  """Returns high confidence keypoints and edges for visualization.
-
-  Args:
-    keypoints_with_scores: A numpy array with shape [1, 1, 17, 3] representing
-      the keypoint coordinates and scores returned from the MoveNet model.
-    height: height of the image in pixels.
-    width: width of the image in pixels.
-    keypoint_threshold: minimum confidence score for a keypoint to be
-      visualized.
-
-  Returns:
-    A (keypoints_xy, edges_xy, edge_colors) containing:
-      * the coordinates of all keypoints of all detected entities;
-      * the coordinates of all skeleton edges of all detected entities;
-      * the colors in which the edges should be plotted.
-  """
-  keypoints_all = []
-  keypoint_edges_all = []
-  edge_colors = []
-  num_instances, _, _, _ = keypoints_with_scores.shape
-  for idx in range(num_instances):
-    kpts_x = keypoints_with_scores[0, idx, :, 1]
-    kpts_y = keypoints_with_scores[0, idx, :, 0]
-    kpts_scores = keypoints_with_scores[0, idx, :, 2]
-    kpts_absolute_xy = np.stack(
-        [width * np.array(kpts_x), height * np.array(kpts_y)], axis=-1)
-    kpts_above_thresh_absolute = kpts_absolute_xy[
-        kpts_scores > keypoint_threshold, :]
-    keypoints_all.append(kpts_above_thresh_absolute)
-
-    for edge_pair, color in KEYPOINT_EDGE_INDS_TO_COLOR.items():
-      if (kpts_scores[edge_pair[0]] > keypoint_threshold and
-          kpts_scores[edge_pair[1]] > keypoint_threshold):
-        x_start = kpts_absolute_xy[edge_pair[0], 0]
-        y_start = kpts_absolute_xy[edge_pair[0], 1]
-        x_end = kpts_absolute_xy[edge_pair[1], 0]
-        y_end = kpts_absolute_xy[edge_pair[1], 1]
-        line_seg = np.array([[x_start, y_start], [x_end, y_end]])
-        keypoint_edges_all.append(line_seg)
-        edge_colors.append(color)
-  if keypoints_all:
-    keypoints_xy = np.concatenate(keypoints_all, axis=0)
-  else:
-    keypoints_xy = np.zeros((0, 17, 2))
-
-  if keypoint_edges_all:
-    edges_xy = np.stack(keypoint_edges_all, axis=0)
-  else:
-    edges_xy = np.zeros((0, 2, 2))
-  return keypoints_xy, edges_xy, edge_colors
- */
-
-/** */
-fun main(): Unit = moveNetRawPrediction()
-
